backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Myfz <fspielmann@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.14.0"
requests = "^2.28.2"
firebase-admin = "^6.1.0"
mysqlclient = "^2.1.1"


[build-system]
requires = ["poetry-core>=1.4.0"]
build-backend = "poetry.core.masonry.api"